+++ src/app/register/store/styles.css.ts
@@ -1,16 +1,29 @@
-import { colors } from '@/styles/constant';
+import { colors, fontWeight } from '@/styles/constant';
 import { style } from '@vanilla-extract/css';
 
-export const styles = {
-  container: style({
-    display: 'flex',
-    flexDirection: 'column',
-    gap: 28,
-  }),
-  primaryText: style({
-    color: colors.primary,
-  }),
-  darkGrayText: style({
-    color: '#5D5D5D',
-  }),
-};
+export const container = style({
+  display: 'flex',
+  flexDirection: 'column',
+  paddingTop: 10,
+  position: 'absolute',
+  width: '100vw',
+  left: 0,
+  alignItems: 'center',
+  gap: 20,
+});
+
+export const image = style({
+  width: '100%',
+  height: 300,
+});
+
+export const message = style({
+  fontSize: 24,
+  lineHeight: 1.6,
+  fontWeight: fontWeight.bold,
+  textAlign: 'center',
+});
+
+export const messageDetail = style({
+  color: colors.lightGray150,
+});
